reorganized file locations
