+++ libs/smart-ngrx/src/actions/action.service/mark-feature-parents-dirty.function.ts
@@ -1,5 +1,5 @@
-import { ParentInfo } from '../../actions/parent-info.interface';
 import { forNext } from '../../common/for-next.function';
+import { ParentInfo } from '../parent-info.interface';
 import { markParentsDirty } from './mark-parents-dirty.function';
 
 /**
