+++ libs/smart-ngrx/src/effects/effects-factory/long-running-observables/manage-maps.function.ts
